+++ datamodel/openapi/openapi-core-apache/src/test/java/com/sap/cloud/sdk/services/openapi/apache/apiclient/ApiClientExtensionsDeserializationTest.java
@@ -1,39 +1,32 @@
-package com.sap.cloud.sdk.services.openapi.apiclient;
+package com.sap.cloud.sdk.services.openapi.apache.apiclient;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.entry;
-import static org.springframework.test.web.client.match.MockRestRequestMatchers.method;
-import static org.springframework.test.web.client.match.MockRestRequestMatchers.requestTo;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Set;
+import java.util.StringJoiner;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.client.ExpectedCount;
-import org.springframework.test.web.client.MockRestServiceServer;
-import org.springframework.test.web.client.response.MockRestResponseCreators;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
 
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.github.tomakehurst.wiremock.client.WireMock;
+import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
 import com.github.tomakehurst.wiremock.junit5.WireMockTest;
-import com.sap.cloud.sdk.services.openapi.core.AbstractOpenApiService;
+import com.sap.cloud.sdk.services.openapi.apache.core.OpenApiRequestException;
 
+import jdk.jfr.Description;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -77,21 +70,17 @@ class ApiClientExtensionsDeserializationTest
         """;
 
     @Test
-    void testDeserializeResponseWithNestedExtensions()
+    @Description( "Tests deserialization of responses with nested extensions using the Apache HTTP client based ApiClient." )
+    void testDeserializeResponseWithNestedExtensionsApache( WireMockRuntimeInfo wmInfo )
     {
-        final ApiClient apiClient = new ApiClient().setBasePath(BASE_PATH);
-        final RestTemplate restTemplate = apiClient.getRestTemplate();
-        final MockRestServiceServer server = MockRestServiceServer.bindTo(restTemplate).build();
 
-        server
-            .expect(ExpectedCount.once(), requestTo(BASE_PATH + RELATIVE_PATH))
-            .andExpect(method(HttpMethod.GET))
-            .andRespond(MockRestResponseCreators.withSuccess(RESPONSE, MediaType.APPLICATION_JSON));
+        WireMock.stubFor(WireMock.get(WireMock.urlEqualTo(RELATIVE_PATH)).willReturn(WireMock.okJson(RESPONSE)));
+        final var apiClient = ApiClient.create().withBasePath(wmInfo.getHttpBaseUrl());
 
-        final TestSpringApi api = new TestSpringApi(apiClient);
+        final TestApacheApi api = new TestApacheApi(apiClient);
         final Outer result = api.getOuter();
 
-        server.verify();
+        WireMock.verify(1, WireMock.getRequestedFor(WireMock.urlEqualTo(RELATIVE_PATH)));
         assertResult(result);
     }
 
@@ -136,46 +125,52 @@ class ApiClientExtensionsDeserializationTest
         });
     }
 
-    private static class TestSpringApi extends AbstractOpenApiService
+    private static class TestApacheApi extends BaseApi
     {
-        public TestSpringApi( final ApiClient apiClient )
+        public TestApacheApi( final ApiClient apiClient )
         {
             super(apiClient);
         }
 
+        @Nonnull
         public Outer getOuter()
+            throws OpenApiRequestException
         {
             final Object localVarPostBody = null;
-            // create path and map variables
-            final Map<String, Object> localVarPathParams = new HashMap<String, Object>();
-            final String localVarPath =
-                UriComponentsBuilder.fromPath("/outer").buildAndExpand(localVarPathParams).toUriString();
 
-            final MultiValueMap<String, String> localVarQueryParams = new LinkedMultiValueMap<String, String>();
-            final HttpHeaders localVarHeaderParams = new HttpHeaders();
-            final MultiValueMap<String, Object> localVarFormParams = new LinkedMultiValueMap<String, Object>();
+            // create path and map variables
+            final String localVarPath = RELATIVE_PATH;
+
+            final StringJoiner localVarQueryStringJoiner = new StringJoiner("&");
+            String localVarQueryParameterBaseName;
+            final List<Pair> localVarQueryParams = new ArrayList<Pair>();
+            final List<Pair> localVarCollectionQueryParams = new ArrayList<Pair>();
+            final Map<String, String> localVarHeaderParams = new HashMap<String, String>();
+            final Map<String, Object> localVarFormParams = new HashMap<String, Object>();
 
             final String[] localVarAccepts = { "application/json" };
-            final List<MediaType> localVarAccept = apiClient.selectHeaderAccept(localVarAccepts);
-            final String[] localVarContentTypes = {};
-            final MediaType localVarContentType = apiClient.selectHeaderContentType(localVarContentTypes);
+            final String localVarAccept = ApiClient.selectHeaderAccept(localVarAccepts);
 
-            final String[] localVarAuthNames = new String[] {};
+            final String[] localVarContentTypes = {
 
-            final ParameterizedTypeReference<Outer> localVarReturnType = new ParameterizedTypeReference<Outer>()
+            };
+            final String localVarContentType = ApiClient.selectHeaderContentType(localVarContentTypes);
+
+            final TypeReference<Outer> localVarReturnType = new TypeReference<Outer>()
             {
             };
             return apiClient
                 .invokeAPI(
                     localVarPath,
-                    HttpMethod.GET,
+                    "GET",
                     localVarQueryParams,
+                    localVarCollectionQueryParams,
+                    localVarQueryStringJoiner.toString(),
                     localVarPostBody,
                     localVarHeaderParams,
                     localVarFormParams,
                     localVarAccept,
                     localVarContentType,
-                    localVarAuthNames,
                     localVarReturnType);
         }
     }