+++ datamodel/openapi/openapi-core-apache/src/test/java/com/sap/cloud/sdk/services/openapi/apache/apiclient/ApiClientFromDestinationTest.java
@@ -1,4 +1,4 @@
-package com.sap.cloud.sdk.services.openapi.apiclient;
+package com.sap.cloud.sdk.services.openapi.apache.apiclient;
 
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -11,7 +11,6 @@ import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import com.sap.cloud.sdk.cloudplatform.connectivity.DefaultHttpDestination;
 import com.sap.cloud.sdk.cloudplatform.connectivity.Destination;
 import com.sap.cloud.sdk.cloudplatform.connectivity.HttpDestination;
-import com.sap.cloud.sdk.services.openapi.core.AbstractOpenApiService;
 
 class ApiClientFromDestinationTest
 {
@@ -24,9 +23,10 @@ class ApiClientFromDestinationTest
     {
         final HttpDestination testDestination = DefaultHttpDestination.builder(SERVER.baseUrl()).build();
 
-        final MyTestAbstractOpenApiService springService = new MyTestAbstractOpenApiService(testDestination);
+        final MyTestAbstractApacheOpenApiService apacheService =
+            new MyTestAbstractApacheOpenApiService(testDestination);
 
-        springService.foo();
+        apacheService.foo();
     }
 
     @Test
@@ -34,15 +34,15 @@ class ApiClientFromDestinationTest
     {
         final HttpDestination testDestination = DefaultHttpDestination.builder(SERVER.baseUrl()).build();
 
-        final MyExceptionThrowingSpringServiceAbstract springService =
-            new MyExceptionThrowingSpringServiceAbstract(testDestination);
+        final MyExceptionThrowingApacheServiceAbstract apacheService =
+            new MyExceptionThrowingApacheServiceAbstract(testDestination);
 
-        assertThatExceptionOfType(IllegalAccessException.class).isThrownBy(springService::foo);
+        assertThatExceptionOfType(IllegalAccessException.class).isThrownBy(apacheService::foo);
     }
 
-    private static class MyTestAbstractOpenApiService extends AbstractOpenApiService
+    private static class MyTestAbstractApacheOpenApiService extends BaseApi
     {
-        public MyTestAbstractOpenApiService( final Destination destination )
+        public MyTestAbstractApacheOpenApiService( final Destination destination )
         {
             super(destination);
         }
@@ -53,9 +53,9 @@ class ApiClientFromDestinationTest
         }
     }
 
-    private static class MyExceptionThrowingSpringServiceAbstract extends AbstractOpenApiService
+    private static class MyExceptionThrowingApacheServiceAbstract extends BaseApi
     {
-        public MyExceptionThrowingSpringServiceAbstract( final Destination destination )
+        public MyExceptionThrowingApacheServiceAbstract( final Destination destination )
         {
             super(destination);
         }